Instructions for building with a vendored OpenSSL
The `openssl-sys` crate expects to find the OpenSSL headers and libraries already installed. If they're not, then installing `cargo-instruments` fails. Installing with the `vendored-openssl` feature works in this case by causing `openssl-sys` to link a static copy of OpenSSL.

The Nix instructions were missing the `openssl` package so I added that too. I think the Nix instructions are wrong over all and should be removed in a subsequent commit. See my comment in #55 for why it's wrong.

Fixes: #72