Upgrade cargo crate
I was trying to use `cargo instruments` on and it was choking on one of my
`Cargo.toml` files.  I'm not sure precisely what it didn't like, but upgrading
the cargo crate has fixed it.  Presumably there's some newer cargo feature that
I'm using and this older crate doesn't support.